# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vinishko AI Sommelier API",
    "description": "Диалог с AI-сомелье по распознанному вину. Состояние сессии хранится в сервисе.",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "service"
        ],
        "summary": "Проверка процесса",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Открыть сессию и получить первое сообщение",
        "operationId": "open_session_v1_sessions__session_id__post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid7",
              "description": "UUID версии 7, созданный клиентом для этой сессии.",
              "title": "Session Id"
            },
            "description": "UUID версии 7, созданный клиентом для этой сессии."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "Получить историю диалога",
        "description": "Возвращает все видимые сообщения сохранённой сессии по порядку: первый ответ сомелье, затем пары «вопрос пользователя — ответ сомелье». У первого ответа есть две подсказки; у остальных сообщений suggestions равно null. Используйте метод для восстановления переписки в интерфейсе после обновления страницы. Для продолжения диалога вызывать его не требуется: POST /v1/sessions/{session_id}/messages сам использует сохранённую историю. Если сессия не найдена, возвращается 404.",
        "operationId": "get_session_v1_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid7",
              "description": "UUID версии 7, созданный клиентом для этой сессии.",
              "title": "Session Id"
            },
            "description": "UUID версии 7, созданный клиентом для этой сессии."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sessions"
        ],
        "summary": "Удалить сессию",
        "operationId": "delete_session_v1_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid7",
              "description": "UUID версии 7, созданный клиентом для этой сессии.",
              "title": "Session Id"
            },
            "description": "UUID версии 7, созданный клиентом для этой сессии."
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}/messages": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Отправить следующий вопрос",
        "operationId": "send_message_v1_sessions__session_id__messages_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid7",
              "description": "UUID версии 7, созданный клиентом для этой сессии.",
              "title": "Session Id"
            },
            "description": "UUID версии 7, созданный клиентом для этой сессии."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserMessageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OpenSessionRequest": {
        "properties": {
          "wine": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "minProperties": 1,
            "title": "Wine",
            "description": "Распознанная основная карточка вина."
          },
          "candidates": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object",
              "minProperties": 1,
              "description": "Каталожная карточка. Неиспользуемые поля допускаются, но игнорируются моделью."
            },
            "type": "array",
            "maxItems": 5,
            "title": "Candidates",
            "description": "До пяти альтернативных каталожных карточек, в порядке релевантности."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "wine",
          "candidates"
        ],
        "title": "OpenSessionRequest"
      },
      "PublicMessage": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "suggestions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestions",
            "description": "Ровно две подсказки только у первого ответа; затем null."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role",
          "content",
          "suggestions"
        ],
        "title": "PublicMessage"
      },
      "SessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid7",
            "title": "Session Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/PublicMessage"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "messages"
        ],
        "title": "SessionResponse"
      },
      "TurnResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid7",
            "title": "Session Id"
          },
          "message": {
            "$ref": "#/components/schemas/PublicMessage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "message"
        ],
        "title": "TurnResponse"
      },
      "UserMessageRequest": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Content",
            "description": "Вопрос пользователя или текст выбранной подсказки."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content"
        ],
        "title": "UserMessageRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}